% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_report.R
\name{calibration_report}
\alias{calibration_report}
\title{Generate a calibration report}
\usage{
calibration_report(filepath, output_folder, report_filename = NULL)
}
\arguments{
\item{filepath}{Full path to a calibration .xls file.}

\item{output_folder}{Full path to the folder in which to render the report PDF.}

\item{report_filename}{NULL to copy the .xls filename, or a string to use in its place.}
}
\value{
Nothing in the workspace; a PDF will be rendered to the output_folder.
}
\description{
Use \code{rmarkdown::render()} to render a calibration report from
a calibration sequence .xls file.
}
\examples{
calibration_report(list.files(directory, full.names = TRUE)[1], output_folder = "path/to/output/folder/")

}
\author{
Matt Harris
}
