% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_parsers.R
\name{parse_cal_formula}
\alias{parse_cal_formula}
\title{Parse the formula of a calibration}
\usage{
parse_cal_formula(parsed_calibration, analyte_name)
}
\arguments{
\item{parsed_calibration}{output of \code{parse_sequence()}, applied to a calibration sequence.}

\item{analyte_name}{character, name of analyte. E.g., "MSA".}
}
\description{
Return the calibration regression formula from a calibration list object
as a formula object to pass to e.g., \code{stats::lm()}
}
