% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_sequence.R
\name{parse_sequence}
\alias{parse_sequence}
\title{Parse a sequence excel file.}
\usage{
parse_sequence(filepath, ion_type, verbose = TRUE)
}
\arguments{
\item{filepath}{The full filepath to a sequence .xls file}

\item{ion_type}{One of either 'Anion' or 'Cation'.}

\item{verbose}{TRUE/FALSE to print status messages.
#'}
}
\value{
A list containing sequence information and data.
}
\description{
Read in and parse a sequence excel file into a nested list of its 
components. The sequence type (Calibration or sample run) is detected automatically
based on the filename. Ion type (Anion or Cation) must be specified.

Ref2
}
\examples{
## Code examples here
sequence <- parse_sequence(filepath = "Drive:/full/path/to/file.xls", ion_type = "Anion")
                           
}
\author{
Matt Harris
}
