% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cal_analyte.R
\name{plot_cal_analyte}
\alias{plot_cal_analyte}
\title{Plot the calibration for a target analyte}
\usage{
plot_cal_analyte(
  parsed_calibration,
  analyte_name,
  axlim_offset_pct = 1,
  point_colour = "black",
  regline_colour = "red",
  inset = "beside"
)
}
\arguments{
\item{parsed_calibration}{output from \code{parse_calibration()}.}

\item{analyte_name}{string matching an ion in the target calibration, e.g., "MSA".}

\item{axlim_offset_pct}{numeric; percentage the axes will be offset from the maximum and minimum value.}

\item{point_colour}{character; fill colour of points in calibration scatterplot.}

\item{regline_colour}{character; colour of regression line}

\item{inset}{one of "beside", "inside" or NULL to plot the lowest 10 percent of values as a separate plot beside or inside the main plot, or not at all.}
}
\value{
a cowplot plotting object containing one or more ggplot2 plotting objects
}
\description{
Plot calibration for a single ion species.
}
\author{
Matt Harris
}
