% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_res_analyte.R
\name{plot_res_analyte}
\alias{plot_res_analyte}
\title{Residual plot of a single analyte around a regression}
\usage{
plot_res_analyte(
  parsed_calibration,
  analyte_name,
  unit = "percent",
  axlim_offset_pct = 1,
  point_colour = "black"
)
}
\arguments{
\item{analyte_name}{Exact character name for analyte. Will be used for explicit character matching. E.g., 'Lithium'.}

\item{unit}{one of either 'area' for raw residuals values in microsiemens per minute or in the same units, but expressed as a percentage of the measured value.}

\item{axlim_offset_pct}{numeric; percentage value to offset limits of the y axis beyond the min and max values. Default is 1.}

\item{point_colour}{Fill colour for points.}

\item{imported_spreadsheet}{spreadsheet imported containing per-level and per-analyte ppb weights and instrument-returned values.}
}
\value{
A ggplot object.
}
\description{
Plot the residuals of a single analyte relative to its versus-known-weight regression in microsiemens per minute (peak area).
}
\author{
Matt Harris
}
